+++ CHANGELOG.md
@@ -1,3 +1,13 @@
 # Changelog
 
 ## 0.*
+
+### next
+
+* Set up the project with basic HTML pages.
+* Added support to WeNet OAuth2 authentication flow.
+* Integrated survey in the app.
+* Created a web service for parsing survey responses.
+* Set up the structure for updating a user profile.
+* Added sentry integration
+* Added periodic task to handle possible errors when updating the profile, every 15 minutes checks if there are errors and tries to re-execute the tasks that causes the errors